hooks: Pass 'ObjectHash' down to 'configureHooks'
The 'configureHooks' function currently calls `DetectObjectHash()` to
obtain the object hash of the repo. This works, but this doesn't utilize
the repository cache present. So instead, pass down the 'ObjectHash' to
'configureHooks'. This should ensure we obtain the object hash from a
'RepositoryExecutor' which has caching abilities.

We don't send the 'RepositoryExecutor' down to 'configureHooks' because
these configurations are often called as part of other Git commands and
our tests use 'CountingCommandFactory' to count the number of commands
being run. Since the lock is obtained on the parent command, trying to
run the 'RepositoryExecutor' inside the hook would cause a deadlock.

This itself isn't reason enough, since we shouldn't write code around
tests. But considering the functionality of this ('configureHooks')
function and its immediate parents, it makes sense for them to not run
any Git commands within them. So taking out the Git commands from them
makes more sense.

Also move the config param right after the context param to make the
arguments a little cleaner.

Re-create the localrepo in the backup package when a new repo is created
so that the cache for object hash is invalidated.