Move LogManager interface to backup package
LogManager is the interface the WAL archiver expects to interact
with. It's not a general storage interface. Move the interface to
the backup package.

The WAL archiver's lifecycle is not tied to the partition's lifecycle.
The TransactionManager may shutdown even if the WAL is not fully archived.
The WAL archiver will start it again when it gets a handle to the partition
through Node. However, this does not expose the LogManager interface's
methods as those do not belong on the general Partition interface. We thus
cast the returned Partition to the LogManager type to get access to the
methods similarly as the previous CallLogManager implementation did.

This could be improved upon later by scoping the LogConsumer passed to
a TransactionManager to a specific partition. Ideally the partition would
also stay up until the log consumer has managed to consumer the log. This
would better match the general behavior with Raft as well, since we'd want
the partition to stay up until it is fully replicated. If this was the case,
the WAL archiver wouldn't need to restart the partition but could respond
always throug the LogManager interface.