localRepo: Set the reference backend when running 'Exec'
If there are commands being run on the `localrepo.Repo` using the 'Exec'
command we want to set the reference backend metric label in the same.

This is done on the `localrepo.Repo` level and not on the `gitcmd`
level, because `localrepo.Repo` provides caching of the reference
backend value. If we're running on the `gitcmd` layer, then we'd have to
call Git each time to get the reference backend and this can be
expensive.

This also goes hand in hand with the recent changes, where we converted
most of our code to use `localrepo.Repo` and propagate that through the
different layers, instead of using `gitcmd` directly.