Don't retain write buffer in SearchFilesByContent
SearchFilesByContent calling ServerStream.Send() directly on the
write buffer passed to it. The buffer comes from io.Copy and may
be reused and modified afterwards. gRPC has documented that it is
not safe to modify the message passed into to Send().

Copy the write buffer instead of passing it directly to Send.