Handle unique conflicts from storage_repositories_new_pkey
Praefect contains two unique indexes on the storage_repositories
table due to a migration changing the primary key to repository_id
which was never finished.

We're currently not handling the unique index conflicts from the
new index, only the earlier one. Add handling for this so the
errors are properly converted to the domain error.

Typically errors would be raised from the first index only as it
gets updated first. There's been reports of the order changing when
logical replication is used.