ff: Remove gitaly_use_resizable_semaphore_in_concurrency_limiter flag
This feature flag swaps in resizable semaphore. This semaphore is the
core of Gitaly's concurrency limiting. It controls the maximum N actors
that can acquire the resource at the same time and can be resized
freely.

This flag has been enabled for around 1 year. It is also enabled by
default. It's time to remove it.

Rollout issue: https://gitlab.com/gitlab-org/gitaly/-/issues/5581