limiter: Remove redundant static semaphore
In a prior commit, we removed a feature flag that controls the type of
semaphore used for the concurrency limiter. The resizable semaphore
is used by default now. It is a more advanced semaphore that supports
dynamic resizing. The alternative static semaphore is not used
anywhere else.

This commit removes that redundant semaphore implementation.