linter: Replace deprecated linter
Go previously had a common problem where reference to a loop variable
was leaking to next iterations. And we had to explicitly copy the loop
variable to fix this issue. With Go 1.22, for loop scoping has changed
and it fixed the mentioned issue. Now we can replace the exportloopref
linter with copyloopvar and get rid of all the reassigns we did before.