golangci: Re-organize the linters
Some of the linters are default enabled in golangci-lint. However we
disable all linters to ensure that there are no surprises whenever we
update golang-ci lint.

Let's re-organize this list however to make it a bit easier to manage.
We first split it into two, one for the default enabled linters and the
second for the rest. Then we sort them alphabetically.