repository: Use `bufio.NewReader` in `SearchFilesByContent`
The `SearchFilesByContent` RPC is used to search for files which contain
a particular pattern. This is just a wrapper around the `git-grep`
command. When files have large lines in them, we currently through a
`scan: bufio.Scanner: token too long` error. This is because we use
`bufio.NewScanner` which doesn't work with large buffer sizes.

To overcome this issue, let's use the suggested `bufio.NewReader`
instead. This solves our large line issue and fixes the broken test
added in the previous commit.