catfile: Remove gitaly-session-id
gitaly-session-id is a unique session that gets injected into the gRPC
context before calls, that we use as part of the cache key for the
catfile process cache. This was originally to achieve separation with
regards to each rails request with the worry that if two rails requests
shared the same cache that one affect the other.

However, the catfile cache is not a in memory cache that stores data per
se, but a process cache that prevents from having to spawn a new git
process for every time a client needs access to a git object.

This means that there is no reason why different requests for the same
repository cannot share the same catfile process cache.

Let's remove the session ID from the cache key behind a feature flag.
Instead, replace it with a key that is the nearest five minute value.
This ensures that the cache entry is recycled at least every five
minutes.