Pass around localrepo instance during reference verification
TransactionManager is creating a second localrepo instance of the
staging repository when verifying reference. This has a performance
impact as the second instance does not share the cached values for
reference backend and object hash.

Pass down the existing instance instead to optimize performance by
benefiting from the caching of these values.