Configure attr.tree=HEAD for performance sensitive commands
ExecCommandFactory is configuring attributes to be read a from
an empty tree by every Git command that doesn't need attributes.
To do so, it needs to resolve the object hash of the repository
to set the correct empty tree object. Doing so, it also doesn't
use the cache to avoid recursion. Resolving the object hash adds
significant overhead to transaction processing, especially as it
is done for commands like update-ref, cat-file and rev-parse that
are executed in the critical section.

Point attributes to HEAD for these commands to avoid the expensive
object hash resolving.