Share cached values with quarantined localrepo instances
Quarantine() applies a quarantine to a given localrepo instance.
As it currently return a new localrepo instance, it loses the cached
values for object hash and reference backend, and must re-resolve
them again. This is impacting the performance of transaction manager.
Share the resolved values between the original and the quarantined
instance of the localrepo.