Resolve reference OIDs with a single process
TransactionManager is currently calling ResolveRevision for each
reference the transaction is updating to figure out the current
value of the reference. This leads to spawning a rev-parse command
for each reference which is very inefficient if a transaction
contains more than one reference.

Use the existing checkObjects function to resolve the old values
of the references using a single command.