Update forking time histogram buckets
Forking time histogram's buckets are not a good fit for the actual
time the operations are expected to take. Use exponential buckets
instead of hand setting the buckets. This will give us a progression
with more buckets in the shorter durations which is where we expect
most observations to land. We cap the buckets at 1s as that is
sufficient to see from the metrics that forking is taking way too
long.