Write and read directly from cat-file
checkObjects is currently using multiple a goroutines to write
input to cat-file, wait for cat-file, and read output from cat-file.
It's using io.Pipe to pass input and output between these goroutines.
This is inefficient. io.Pipe is a synchronous in-memory structure.
Passing an io.Pipe as stdin to the command leads it to launch a
goroutine that copies from the in-memory pipe to the actual stdin
pipe of the command.

Stop using io.Pipe and instead write directly to the pipe that is
connected to the cat-file command's stdin. We do this by using
WithSetupStdin that leads to `os/exec` setting up an `os.Pipe` and
passing that directly to the command. Doing so, we can also reduce
the number of goroutines used as there's no need for separate
goroutine to copy the input written to the pipe to the command.
We also move the reader to the main goroutine as it would anyway wait
for the output to be read.