raft: Remove dragonboat-based Raft package
In the latest decision, we would like to pivot the etc/raft library as a
replacement for dragonboat. Those two libraries are absolutely different
in terms of philosophy and design. Also, as the project switched the
direction to bottom-up, we'll start by injecting etcd/raft directly into
the WAL transaction manager. Thus, it makes a lot of sense to remove
components built based on dragonboat at the moment. It doesn't make any
sense to keep two of them co-exist at the same time. Of course, we don't
want to throw everything away.

The history is preserved in Git so that it's trivial to cherry-pick
necessary components from the history later.