Read object hash directly from configuration
Gitaly is currently using a Git command to determine the object
hash of the repository. This is incredibly inefficient for such a
simple operation. This commit reads the object hash directly from
the repository's configuration file to optimize.

While we could, we don't fully parse the configuration file or
validate its structure. Gitaly is responsible for writing the
configuration files so we can ensure we write well formed
configuration files.