Remove redundant fsyncs
Begin used to be the place where a partition was started. As it was
responsible for setting up a partition, it created its state
directory and fsynced it. Partition starting logic has since been
moved to startPartition, which also creates the partition's state
directory and fsyncs it. Doing this twice is redundant, and affects
performance with file systems that flush out more writes than just
the fsynced file itself. Remove the redundant logic.