storagemgr: Fix partition id extraction
Format of the key we are trying to extract partition id 
is "p/{big endian represented partition id}/applied_lsn".
Previously we were parsing this key with forward slashes
and use the 2nd element as partition id. But big endian
representation of a valid partition id can also contain
forward slash which results in broken partition id.

Instead we will change this logic to read 8 bytes after
"p/" prefix. Which is guaranteed to be a valid partition id.