Add repo_changed field in FetchRemoteResponse RPC
This commit adds the repo_changed field in FetchRemoteResponse.
If the repository changes, this field is set to true.

The goal of this field is to allow the GitLab project to begin
syncing LFS objects before we update branches in that project.
This should avoid a known race condition when cloning/fetching
repositories and LFS objects are not synced.

If the repository changes successfully, we are ready to perform
the Git operations.

Changelog: added