Start partitions concurrently
We were previously fsyncing outside of the lock block in startPartition.
This led to fsyncing on every request. We're now only fsyncing when
setting up the partition but this leads to doing a potentially longer
running operations in the lock block. Holding the lock while the fsync
is running blocks other unrelated partitions from being retrieved
concurrently.

This commit adds additional synchronization to ensure multiple partitions
can be started concurrently. The partitions start in uninitialized state.
The goroutine starting the partition signals to the other goroutines waiting
for the partition to start when it has finished setting it up. This allows
it to release the lock while setting up the state directories and fsyncing
them.