cgroups: Provide existing manager to ExecCommandFactory
We initialise a cgroup manager during server initialisation, but we
don't provide it to the constructor of the ExecCommandFactory. This
causes us to initialise two instances of the cgroup manager; once in the
server initialisation, and again within the factory's constructor.

Creating cgroup managers is an idempotent operation, which is why it's
been working fine as-is, but it does waste a lot of cycles. Fix this by
providing the outer cgroup manager to the factory constructor.