check: Ensure unpacked binaries are cleaned up
The `gitaly check` command instantiates a GitLabHookManager to verify
that Gitaly is able to access Rails' /internal/check endpoint. The hook
manager requires an ExecCommandFactory which requires Git binaries to be
configured properly. Due to this, we must unpack the embedded Git
binaries into `cfg.RuntimeDir` when this command is executed.

There were three problems with the implementation:

- We didn't clean up the runtime directory, so multiple invocations of
  the command would fail when it tries to unpack binaries into an
  existing location.
- We use `cfg.RuntimeDir` directly as configured in config.toml. We
  should instead be creating subdirectories like we do elsewhere.
- We didn't clean up leftover runtime directories, which may exist if
  a previous process crashed.

Address these problems by unpacking binaries into a subdirectory of
`cfg.RuntimeDir`, and ensure the subdirectory is cleaned up at the
beginning and end of execution. We use `config.SetupRuntimeDirectory()`
which is the same mechanism used by the sever initialisation.