Use unique objects for every reference update
BenchmarkTransactionManager is currently using the same commit
objects for every reference update targeting a given repository.
This hampers benchmarking as we'd like to also measure how long it
takes to verify all of the dependency objects exist in the repository.
We'd want to use unique objects to ensure the number of dependency
objects also scales with the transaction size.

This commit changes all of the reference updates to use a unique
object ID.