Benchmark transactions per second
BenchmarkTransactionManager is currently measuring the number of
reference updates per second. This is a leftover from the time the
benchmark was introduced as transaction manager only supported
reference updates. Nowadays it can also write objects, and the
transactions may not have any reference updates for example if a
repository is being deleted as those would be plain file operations.

Measure the performance on a more general level by recording the
transactions per second rather than reference updates per second.