storagemgr: Re-introduce intermediate appendedLSN index
In the previous commit, the TransactionManager renames appendedLSN to
committedLSN. This commit re-introduces appendedLSN but with a different
meaning. It now tracks the intermediate state waiting to be committed by
other nodes.

The `commitLogEntry` is split into two smaller functions. The first one
encapsulates and updates appendedLSN. The second one is supposed to
engage Raft updates committedLSN. As now we haven't integrated the Raft
library yet, the second function simply increases the committedLSN
index.

This commit also introduces a new non-subtle change: committedLSN does
not persist in KV DB but is next to appliedLSN. Before, as soon as a log
entry is flushed to disk, that log entry will be applied, eventually.
After splitting the append stage and commit stage, things might go wrong
after a log entry is appended but is not committed. This is particularly
true if Raft enters the picture and adds network latency to the log
entry proposal.

Thus, the transaction manager needs to persist the committedLSN to avoid
applying for uncommitted entries after a crash. After restarts, it
resumes from the latest committedLSN, applies unapplied entries, if any,
and removes obsoleted appended entries, if any.