raft: Remove cluster.proto
This commit removes cluster.proto. It contains protobuf definitions for
Raft. After we got rid of dragonboat in favor of etcd/raft, they became
obsolete. We'll introduce a new set of alternative protobuf.

The debug tool `gitaly-db` is impacted by this change. Fortunately, it
uses removing protobufs for formatting only.