git: Make Git v2.47 the default
Now that we have two versions of Git, let's make the newer version the
default. This allows us to remove the older version in a follow up
commit.

Also remove the requirement in the reftable CI job, since we already
meet the required version. This bumps the reftable tests to run with Git
v2.47, which included changes to allow reftables to compact tables, even
if some tables were locked. This requries changes in our tests to
reflect the same.