featureflag: Default enable Git v2.47
We slowly rolled out the new Git version on production. Now that the
rollout is completed, we can default enable the new Git version. This is
an intermediate step before we make it the default version, which we
will do in the following release cycle. While it is no longer required
to rollout new Git version over multiple releases, we take a more
cautious approach, to ensure there are no performance regressions and if
there are, we have options to revert.