Keep idle partitions on standby
StorageManager is currently closing partition's after there are no
more active handles to them. While this cleans up the resources
immediately, it can lead to thrashing under workloads where there's
no concurrency but lots of serial requests.

This commit solves the thrashing problem by keeping a limited number
of inactive partitions on standby. This keeps the partition running
and initialized so it is immediately ready for a follow up request
with no initialization overhead. Once the set limit of inactive
partitions on standby is reached, the least recently used inactive
partition is evicted and cleaned up.

This enables active partitions to remain ready to go while limiting
still the number of partitions we keep open.