Rename activePartitions to runningPartitionGoroutines
Rename the field to a more descriptive name. The WaitGroup is
tracks how many goroutines are still running that were launched
by the StorageManager to run partitions. While the goroutine may
be running, the partition is not necessarily alive anymore as the
goroutine might be performing clean up after the partition has
finished.