Keep unused shared snapshots on standby
SnapshotManager is currently closing shared snapshots as soon as they
have no users. While this clean up releases resources associated with
the snapshots as soon as possible, it leads to thrashing if there are
multiple sequential non-overlapping read operations targeting the same
repository. As soon as the first operation finishes, the snapshot is
removed but needs to be recreated immediately when a subsequent read
arrives.

This commit improves the logic by keeping a limited number of shared
snapshots cached so follow up reads can immediately access them without
the overhead of creating another snapshot and tearing it down. Shared
snapshots are closed if:
- Cache limit is exceeded. If so, the least recently used snapshot is
  evicted and cleaned up.
- The partition is written to. Read snapshots are invalidated to ensure
  the latest data is always read.
- The partition closes leading to releasing all resources associated
  with it. As we keep inactive partitions on standby as well, typically
  we'd only tear down partitions after the maximum number of standbys
  is exceeded, and even then we'd tear down the least recently used
  inactive partition.

Shared snapshots that are invalidated by writes are closed in the
background by a pool of workers to ensure TransactionManager doesn't
block on closing a potentially large number of snapshots.