Perform writes through the API in tests
Gitaly has a number of tests that write directly to the repository.
These will soon break as Gitaly will start caching read snapshots.
Direct writes that don't go through the API don't invalidate the
read snapshots, and the tests wouldn't see their modifications on
subsequent reads. Modify the tests to perform their writes through
the API.

RepositorySize is writing garbage in a manner that we don't have a
good API endpoint to perform the write. Solve the problem there by
invalidating the read snapshots with a no-op write.