Force exclusive snapshots in recovery middleware
The recovery middleware is enabled when transactions have been disabled
but they have been enabled in the past. It ensures that WAL is fully
applied before allowing non-transactional access to the repository. This
ensures the repositories are not in bad states from partially applied WAL.
It achieves this by simply starting a transaction and rolling it back as
this guarantees all transactions have been applied. As we're now keeping
read snapshots open for a potentially long time, these recovery
transactions could stay open and waste disk space.

We've previously enabled transactions on production canary, and this would
potentially impact that node. Force exclusive snapshots to be used for the
transactions so they are discarded immediately when rolling back the
transaction.