Don't keep inactive partitions active in recovery mode
If transactions have been disabled after being enabled, a recovery
middleware is installed to ensure the WAL is fully applied before
allowing access to the repositories in a non-transactional manner
again. If we are in recovery mode, we don't want to keep partitions
on standby as they would just sit idle after the WAL recovery has
been performed. The inactive partition cache can't be turned off, so
set it to a low value so we don't keep lots of partitions active
unnecessarily.