update-ref: Make using 'symref-update' default behavior
We added a featureflag to rollout using 'symref-update' in
'git-update-ref(1)'. The rollout is now done, so we can remove the flag
and cleanup all code.

This involves cleaning the old flow where we manually obtained a lock on
the 'HEAD' file and instead now relying completely on the reference
transaction hook.

Also, this requires fixing up tests as needed.