storagemgr: Centralize errSimulateCrash emitter
The test suite of storage inserts a bunch of test hooks that panic with
errSimulateCrash. Those hooks are unconditional. This commit creates a
helper function that returns a hook function that emits the panic error.
This is a preparation step for later commits where the crashing occurs
for certain entries only.