raft: Implement event registry
etcd/raft provides Propose interface to submit a log entry to the
cluster. This interface exits when the Raft engine accepts the entry.
There is no guarantee when the proposal is committed by a quorum or even
if the proposal ever leaves the node at all.

In some later commits, the transaction manager must propose WAL log
entries in the process loop. Without concrete evidence of log entry
being committed, it's risky to continue. Thus, this commit implements an
event registry. Before the caller proposes, it announces this proposal
to the event registry and embeds the returned event ID in the
transporting Raft message. After the log entry is committed, the caller
is unlocked.