Remove ClockSync check
ClockSync was a diagnostic check to ensure the clocks of the nodes
were in acceptable sync. It was added back in the day as clock drift
could cause authentication issues which weren't immediately obvious.
As we've now fixed the auth errors to actually describe the problem,
remove the clock sync check as unnecessary.