partition: Modify BackupPartition response
Return relative backup path in the BackupPartition RPC response
which will be used to create backup manifest.
With this, we also need to change the way we handle the already
exists backup case. We count them as successful backups in the
PartitionBackupManager, but while it is returning error, it can't
return the backup relative path response. Therefore I am changing
it's implementation to not return anymore and have it a successful
flow by default.