Sync WAL entry's files concurrently
TransactionManager unfortunately has to sync the WAL entry's contents
before appending to log as the reference files and the manifest are
only written out during the critical section. Currently we flush the
files in the WAL entry sequentially. This potentially adds a lot of
overhead as each sync call blocks before the next one is issued.
Improve the latency by issuing fsyncing the WAL entry's files
concurrently.

We use an arbitrary limit of 25 as a safe guard to limit the number of
goroutines launched. Most transactions should anyway be under this but
particularly large fetches and such could still exceed this.