partition: Wire migration manager
The `partition.Factory` is used to configure a `storagemgr.Partition`.
The previously introduced `migration.migrationManager` wraps
`partition.TransactionManager` to provide migration handling for
repositories in a partition.

Introduce a new `migration.migrationFactory` that wraps a
`storagemrg.PartitionFactory` in a similar fashion. Update call sites to
leverage this new factory to wire usage of the migration manager.