Add helper for deciding whether an fsync is needed
Gitaly needs to fsync writes in multiple locations when it isn't
running with transactions enabled. With transactions enabled,
these flushes are unnecessary and hurt performance. The flushes
would target the files in the snapshot which are temporary files.
TransactionManager would flush the files going into the WAl as
needed.

Add a helper to determine whether other parts of the code base
need to flush by checking whether there is a transaction in the
context.