Stop fsyncing in handlers when transactions are enabled
Gitaly is flushing out files in multiple locations. With transactions,
these flushes are unnecessary as they target files in the transaction
snapshots. Stop flushing if we're running in context of a transaction,
and leave the flushing to transaction manager. Unnecessary flushing
hurts performance.

Some function signature are adapted to pass down the context so we
can further pass it down to the check.