Move metadata file functionality to its own package
The storage metadata file is a legacy remain from the time when
Rails was accessing repositories directly on the disk. It should
be removed but there are some references still in Rails.

It creates an import cycle between the safe and storage packages
if storage is references from safe. Move it into its own package
so we can import a utility from the storage package we'll soon
need.