protobuf: Add delete_all to DeleteRefs
Currently DeleteRefs does not contain option to delete all refs.
We are achieving this result by passing HEAD or a non existing prefix
in the except_with_prefix field. But this solution is not ideal. We
should have a way to explicitly say that we want to delete all refs.
Therefore, we are introducing delete_all in the DeleteRefs RPC.