ref: Implement DeleteAll option
We added new delete_all field to the DeleteRefs protobuf. We should
adjust DeleteRefs server implementation to support this new field.
We don't really need to change the actual implementation as it will
be covered by default, but we need to adjust the request validation
to contain the new field.