raft: Implement noop transportation
Introduce a NoopTransport in the Raft module for testing purposes.
This transport logs messages and optionally records them without
performing any actual data transmission. This Transport is used at this
stage so that Raft Manager can swallow all outgoing messages as the
current primary member is the only member of the Raft group.