commit: Return changed files count in CommitStats
The CommitStats RPC provides the number of lines added and deleted for
the requested commit-ish. This is done through executing `git diff
--numstat <commit-ish>`. Each line of output from this command indicates
a changed file along with the number of lines added and deleted. The RPC
iterates through each line and totals the added and deleted lines and
returns the output.

We now want to begin including the number of changed files in the
response as well. This is accomplished by counting each line of Git
output processed and updating the RPC response proto message to include
the file count.

Note that modified binary files are included in the file count, but
omitted from the line counts.