commit: Adjust CommitStats "binary file" test case
The CommitStats RPC has a test case that exercises the RPC when a binary
file is modified in a commit. The commit used for the test case is a
root commit and has no parents. Thus this is really not exercising when
a binary file is changed, but instead when new files are added. This
appears to be an unintential mistake.

Update the test setup to make the binary change commit use the initial
commit as its parent. Also update the corresponding test case to reflect
that no lines were added or deleted since binary files are not included
in line counts.