template: Update Git rollout issue template
Previously we used to allow Git versions to be rolled out in a single
release and did not require a Git version feature flag to be default
enabled before being removed. The feature flag provides a useful escape
hatch if a regression is identified in Git as the flag can be easily
disabled.

Update the Git rollout issue template to now require that at least a
single release has the Git version feature flag default enabled. This
provides more time for issues to be identified across more repositories
while still providing a way to revert back if issues are discovered.