backup: Remove pointer receiver for StorageServiceSink
The only method that modified the receiver was Close and this was trying
to nil out the bucket pointer. This only makes nil pointer exceptions
more likely.

Since the only attribute on this struct is now an unmodified pointer we
can use a non-pointer receiver for all methods. This will mean we aren't
forced to use pointers once we remove the Sink interface.