storagemgr: Rename appendedLSN to committedLSN
Recently, the TransactionManager uses two indices which are appendedLSN
and appliedLSN. It increases the value of appendedLSN after it accepts a
transaction. When the log entry is applied, the transaction manager
increases appliedLSN accordingly. Any log entries prior to appliedLSN
could be removed if there are no pending transactions referring to them.
We are working on adding Raft consensus to the TransactionManager. When
a log entry is accepted by the current node, it also needs to be
transferred to and accepted by the majority of cluster members.
To make it easier for later steps, the TransactionManager needs to track
two indices independently:

* appendedLSN: this index tracks the LSN of the log entry accepted by
the current node, but not yet accepted by others. The transaction
manager are allowed to override entries from committedLSN + 1 to
appendedLSN.
* committedLSN: this index tracks the committed log entries. They are
safe to be applied to the repositories. Eventually, other nodes apply
those entries in the same order.

As the current appendedLSN index is taking the role of committedLSN,
this commit renames it to commitedLSN.