Integrate in-memory reference conflict checks
In the previous commits, we've implemented functionality to conflict
check reference changes against other concurrent changes. This commit
integrates the functionality into TransactionManager.

Reference changes made by a transaction are recorded in the history.
Reference changes being committed subsequently by concurrently
executing transactions are conflict checked against the writes of
earlier transactions. Each transaction is expected to be correct in
isolation so only conflicts introduced by concurrent changes are
detected. The disk state is not looked at. Reference changes are
dropped from the history once all write transactions are executing
with read snapshots that include the changes. That means once all
writes have a read snapshot taken at an LSN equal or later than the
LSN the reference changes were committed under.

This is the first step toward removing 'update-ref' from the critical
section of the TransactionManager. It's the main slow down in verifying
and committing transactions. The conflict checks performed should
already be sufficient to remove 'update-ref' and we don't rely on it
anymore to do any of the checks. However, it's still used for writing
out the reference files and it's file system modifications are recorded
to figure out the exact changes to log. Once we implement file level
conflict checks as well, we'll figure out the file system changes to
perform before entering the critical section, and then just conflict
check them against other transactions.

We already manage to remove the usage of 'cat-file' for figuring out
the old values of references. With a transaction performing a single
reference update, the new code is roughly 434x faster as it goes from
1.7ms to 4μs.

One test is removed as it exercises an unrealistic scenario. The
initial values recorded in the transaction should not be incorrect
as they are recorded in the reference transaction hook before making
the changes. Such a transaction would be invalid. Transactions are
expected to be correct in isolation, and the reference conflict
checks only detect conflicts due to concurrent changes. As this is
not such a conflict, the invalid reference change would not be
detected.